centralize all States and AlertType
