+++ FrostBase/Dtos/State/StateDto.cs
@@ -1,0 +1,10 @@
+public class StateDto
+{
+    public string Id { get; set; }
+    public string Description { get; set; }
+
+    public StateDto FromModel(State s)
+    {
+        return new StateDto();
+    }
+}
