--- FrostBase/Dtos/Truck/StateTruckDto.cs
@@ -1,5 +1,0 @@
-public class StateTruckDto
-{
-    public string Id { get; set; }
-    public string Description { get; set; }
-}
